Version 2.0 Added several things. Details would be illustrated in emails

+++ Controller.py
@@ -1,5 +1,6 @@
 import pygame as pg
 from QuestionBox import QuestionBox as QB
+from PopUpQuestionBox import PopUpQuestionBox as PopUp
 from QuestionGenerator import QuestionGenerator as QG
 from MenuButton import MenuButton
 from MenuButton_ import MenuButton_ as MB
@@ -8,6 +9,7 @@ from ScoreBoard import ScoreBoard as SB
 from HealthIcon import HealthIcon as HI
 import random
 from os import path
+from time import time
 
 
 class Controller:
@@ -16,16 +18,18 @@ class Controller:
         pg.init()
         pg.mixer.init()
         self.screen = pg.display.set_mode((1024, 768))
-        pg.display.set_caption('v2.0')
+        pg.display.set_caption('v2.1')
         self.mouse_x, self.mouse_y = None, None
 
         # some users' identities are created here
         self.questions = pg.sprite.Group()
+        self.pop_up = None
+        self.timer = []
         self.count = 0 # currently has no use
         self.lives = 3
         self.score = 0
-        self.density = 480 # it means problems will be generated every 480 frames
-        self.speed = 3 # the speed of question boxes
+        self.density = 240 # it means problems will be generated every 480 frames
+        self.speed = 1 # the speed of question boxes
 
         # some buttons are created here
         self.start_button = MB((512,404), 'startbasic.png', 'starthover.png')
@@ -57,10 +61,19 @@ class Controller:
             self.sound_effect[i].set_volume(0.15)
 
         # background image
+        self.bg_start_screen_index = 0
+        self.bg_start_screen = [pg.image.load(path.join(base_path, 'assets', 'Background', 'earth_start_screen', 'frame_{}_delay-0.1s.png'.format(x))) for x in range(39)]
+
         self.bg_by_country = {
-            'China': 'background_china.png', (255,255,255)
-            'Egypt': 'background_egypt.png', (222,211,140)
-            'Italy': 'background_italy.png' (179,214,110)
+            'China': pg.image.load(path.join(base_path, 'assets', 'Background', 'background_china.png')),
+            'Egypt': pg.image.load(path.join(base_path, 'assets', 'Background', 'background_egypt.png')),
+            'Italy': pg.image.load(path.join(base_path, 'assets', 'Background', 'background_italy.png'))
+        }
+
+        self.icons_by_country = {
+            'China': ['China.png', 'China_bot.png'],
+            'Egypt': ['Egypt.png', 'Egypt_bot.png'],
+            'Italy': ['Italy.png', 'Italy_bot.png']
         }
 
         # background music
@@ -72,7 +85,7 @@ class Controller:
             'Italy': [None]
         }
 
-    def createQuestionBox(self):
+    def createQuestionBox(self, country):
         '''
             this method creates a questions box object and put it in a sprite group.
         '''
@@ -81,7 +94,7 @@ class Controller:
 
         if self.score <= 10:
             problem = problem_obj.level_1()
-            self.questions.add(QB(x, self.speed, problem[0], problem[1]))
+            self.questions.add(QB(x, self.speed, self.icons_by_country[country][0], problem[0], problem[1]))
             print(problem)
         else: self.questions.add(QB(x, self.speed))
 
@@ -103,7 +116,10 @@ class Controller:
         '''
             show the start button on the screen
         '''
-        self.screen.fill((135,206,250)) # sky blue
+        self.screen.fill((135,206,250))
+        self.screen.blit(self.bg_start_screen[self.bg_start_screen_index//10], self.screen.get_rect()) # draw the earth gif; notice that it's slowed here
+        self.bg_start_screen_index += + 1 if self.bg_start_screen_index < 389 else -388
+
         for button in [self.start_button, self.rules_button, self.credit_button]:
             if self.isOver(button.rect):
                 button.isOver()
@@ -152,34 +168,57 @@ class Controller:
         '''
             start dropping boxes down
         '''
-        self.screen.fill(self.bg_by_country[country])
-        if density == self.density: # the bigger the density is, the slower the game goes
-            self.createQuestionBox()
-            density = 0
-        density += 1
+        if self.score in (3,): # do a pop-up question and freeze the dropping questions
+            self.pop_up = PopUp(self.score)
+            self.timer.append(time())
+            self.screen.blit(self.pop_up.image, self.pop_up.rect)
+            if time() - self.timer[0] >= 5: # if users cannot answer the pop-up question in a certain amount of time, they would lose their chance
+                self.timer = []
+                self.pop_up = None
+                self.score += 0.01
+        else:
+            self.screen.blit(self.bg_by_country[country], self.screen.get_rect())
+            if density == self.density: # the bigger the density is, the slower the game goes
+                self.createQuestionBox(country)
+                density = 0
+            density += 1
 
-        for sp in self.questions:
-            sp.filename = 'moon_2.png'
-            break
-        self.questions.update()
-        self.questions.draw(self.screen)
-        self.deleteOutscreenBox()
+            for sp in self.questions:
+                sp.filename = self.icons_by_country[country][1]
+                break
+            self.questions.update()
+            self.questions.draw(self.screen)
+            self.deleteOutscreenBox()
         return density
 
     def checkAns(self, ans_submitted):
         '''
             check answer submitted by users
         '''
-        for sp in self.questions:
-            if sp.answer == ans_submitted:
+        if self.pop_up is not None:
+            if self.pop_up.problems[self.score]['ans'] == ans_submitted:
                 self.sound_effect['right'].play()
-                self.score_board.update() # in update(), the score will + 1
-                self.score += 1 # I create another instance variable (score) here because it's more convenient
-                self.questions.remove(sp)
+                if self.lives != 3:
+                    self.health_bar.update(1)
+                    self.lives += 1
+                    self.score += 0.01 # see line 171
+                else:
+                    self.score_board.update(5)
+                    self.score += 5
+                self.pop_up = None
             else:
                 self.sound_effect['wrong'].play()
-            break
-        #print(self.score_board.score, self.lives)
+        else:
+            for sp in self.questions:
+                if sp.answer == ans_submitted:
+                    self.sound_effect['right'].play()
+                    self.score_board.update() # in update(), the score will + 1
+                    self.score = int((self.score + 1)//1) # see line 199
+                    self.questions.remove(sp)
+                else:
+                    self.sound_effect['wrong'].play()
+                break
+        print(self.score, self.lives)
 
 
     def startLoop(self):
@@ -235,7 +274,6 @@ class Controller:
         self.screen.blit(self.ans_typein.bg_image, self.ans_typein.bg_rect) # draw the background before drawing the text that users put in
         self.ans.draw(self.screen) # draw the text that users type in
 
-        self.screen.blits(((self.score_board.image, self.score_board.rect), (self.health_bar.byCountry(country), self.health_bar.rect)))
 
         for event in pg.event.get():
             if event.type == pg.QUIT:
@@ -243,12 +281,12 @@ class Controller:
             elif event.type == pg.KEYDOWN and event.key == pg.K_ESCAPE:
                 self.STATE = 'menu'
             elif event.type == pg.KEYDOWN:
-                key_num = event.key # event.key is a number corresponds to the key
-                self.ans.update(key_num)
+                self.ans.update(event)
 
                 ans_submitted = self.ans_typein.submit()
                 if ans_submitted is None: pass
                 else: self.checkAns(ans_submitted) # ans_submiited will not be None if users hit ENTER key with numbers typed in
+        self.screen.blits(((self.score_board.image, self.score_board.rect), (self.health_bar.byCountry(country), self.health_bar.rect)))
         return density
 
     def endLoop(self):
@@ -264,7 +302,7 @@ class Controller:
     def mainloop(self):
         # loop it, or say, run it
         while True: # without this loop, the game will exit automatically after clicking 'start again', because there is no codes after while STATE == 'end' loop
-            density = 360
+            density = 0
             self.volume = 0
             while self.STATE == 'start':
                 self.startLoop()
@@ -294,7 +332,7 @@ class Controller:
             # reset everything
             self.questions.empty()
             self.ans_typein.result = ''
-            self.ans_typein.update(0) # empty the type in box
+            self.ans_typein.update() # empty the type in box
             self.score, self.score_board.score = 0, -1
             self.score_board.update() # have to update to change the image
             self.lives, self.health_bar.health = 3, 4
